feat: #557 fix tests
